feat(ui): refactor role badge handling and improve organizational role display
