+++ client/src/lib/components/members/MemberRow.svelte
@@ -13,6 +13,7 @@
 	import type { OrganizationMembershipResponse, OrganizationalRole } from '$lib/api/models';
 	import { getInitials } from '$lib/utils/string';
 	import { formatRelativeDate } from '$lib/utils/date';
+	import { getOrganizationalRoleBadgeClass } from '$lib/utils/role';
 	import RoleBadge from './RoleBadge.svelte';
 
 	interface Props {
@@ -32,18 +33,6 @@
 		onRemove,
 		onTransferOwnership
 	}: Props = $props();
-
-	function getRoleBadgeClass(role: OrganizationalRole): string {
-		switch (role) {
-			case 'OWNER':
-				return 'bg-gradient-to-r from-purple-500 to-purple-600 border-0 text-white';
-			case 'ADMIN':
-				return 'bg-blue-500/10 border-blue-500/50 text-blue-500';
-			case 'MEMBER':
-			default:
-				return 'border-border/50 bg-background/50 text-muted-foreground';
-		}
-	}
 </script>
 
 <div
@@ -89,7 +78,7 @@
 							size="sm"
 							class="bg-background/50 border-border/50 hover:bg-accent/10"
 						>
-						<Badge class={getRoleBadgeClass(member.role ?? 'MEMBER')}>
+						<Badge class={getOrganizationalRoleBadgeClass(member.role ?? 'MEMBER')}>
 							{#if member.role === 'ADMIN'}
 								<Shield class="mr-1 h-3 w-3" />
 							{/if}
